refactored /txs to rpc.go, http.go and refactored http requests
